added force option

+++ lib/ogre/delete-org.rb
@@ -10,14 +10,14 @@ module Ogre
     argument :server_url , type: :string, desc: DESC_CHEF_SERVER_URL
     argument :key_path, type: :string, desc: DESC_PRIVATE_KEY
 
-    class_option :force, :aliases => '-f', type: :string, desc: DESC_FORCE
+    class_option :force, :aliases => '-f', :default => false, type: :boolean, desc: DESC_FORCE
 
       def delete_org
         # define REST object
         chef_rest = Chef::REST.new(server_url, RUN_AS_USER, key_path)
 
         # prompt user
-        exit unless HighLine.agree("Deleting '#{org}' is permanent. Do you want to proceed? (y/n)")
+        exit unless options[:force] || HighLine.agree("Deleting '#{org}' is permanent. Do you want to proceed? (y/n)")
 
         begin
           chef_rest.delete_rest("/organizations/test")
